requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
groq
langgraph
sentence-transformers
chromadb
python-docx
python-pptx
pymupdf
rank_bm25
pydantic
aiofiles
python-multipart
mistralai
docx2txt
python-docx
python-jose[cryptography]
passlib[bcrypt]
aiohttp
dnspython

